+++ src/main/java/com/ironcompass/training/BankedGoalStatus.java
@@ -1,0 +1,11 @@
+package com.ironcompass.training;
+
+public enum BankedGoalStatus
+{
+    UNKNOWN,
+    NOT_SUPPORTED,
+    NO_RESOURCES,
+    IN_PROGRESS,
+    READY,
+    COMPLETE
+}